Moved ObfuscationOptions into ObfuscationService.
Added Obfuscate function to apply obfuscation steps.
