--- app/Hutch.Relay/Config/ObfuscationOptions.cs
@@ -1,7 +1,0 @@
-namespace Hutch.Relay.Config;
-
-public class ObfuscationOptions
-{
-  public int LowNumberSuppressionThreshold { get; set; }
-  public int RoundingTarget { get; set; }
-}
