+++ app/Hutch.Relay/appsettings.Development.json
@@ -14,5 +14,16 @@
   },
   "Database": {
     "ApplyMigrationsOnStartup": false
+  },
+  "ObfuscationOptions": {
+    "LowNumberSuppressionThreshold": 5,
+    "RoundingTarget": 10
+  },
+  "UpstreamTaskApi": {
+    "BaseUrl": "https://my-task-api.com",
+    "CollectionId": "dd52026b-5a61-4c89-b733-04ba141b3f31",
+    "Username": "user",
+    "Password": "password"
   }
+
 }
